+++ src/web/GoNorth.tsx
@@ -1,5 +1,4 @@
 import { createTheme } from "@mui/material";
-import { deepPurple, pink, purple, teal } from "@mui/material/colors";
 import { ThemeProvider } from "@mui/material/styles";
 import React from "react";
 import { Provider } from "react-redux";